server/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
langchain
openai
python-dotenv
bcrypt==4.0.1
python-jose[cryptography] 
passlib
python-multipart
apscheduler
qdrant-client
pgvector